pyproject.toml
[project]
name = "ygo-judge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.0",
    "googletrans>=4.0.2",
    "langchain-community>=0.3.14",
    "langchain>=0.3.14",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "tqdm>=4.67.0",
    "unstructured>=0.16.13",
    "nltk>=3.9.1",
    "langchain-openai>=0.3.0",
    "chromadb>=0.6.3",
]
